requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-motor==3.4.13
adafruit-circuitpython-motorkit==1.6.14
adafruit-circuitpython-pca9685==3.4.15
adafruit-circuitpython-register==1.10.0
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
evdev==1.7.1
pillow==10.4.0
pyftdi==0.55.4
pynput==1.7.7
pyserial==3.5
python-xlib==0.33
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
sysv-ipc==1.1.0
tk==0.1.0
ttkbootstrap==1.10.1
typing_extensions==4.12.2
